+++ src/layers/cip/core/object.js
@@ -9,75 +9,107 @@ const { DataType } = require('./datatypes/types');
 const { Decode } = require('./datatypes/decoding');
 
 
-const CommonClassAttribute = Object.freeze({
-  Revision: new CIPAttribute.Class(1, 'Revision', DataType.UINT),
-  MaxInstance: new CIPAttribute.Class(2, 'Max Instance ID', DataType.UINT),
-  NumberOfInstances: new CIPAttribute.Class(3, 'Number Of Instances', DataType.UINT),
-  OptionalAttributeList: new CIPAttribute.Class(4, 'Optional Attribute List', DataType.TRANSFORM(
-    DataType.STRUCT([
-      DataType.UINT,
-      DataType.PLACEHOLDER(length => DataType.ABBREV_ARRAY(DataType.UINT, length))
-    ], function (members, dt) {
-      if (members.length === 1) {
-        return dt.resolve(members[0]);
-      }
-    }),
-    value => value[1]
-  )),
-  OptionalServiceList: new CIPAttribute.Class(4, 'Optional Service List', DataType.TRANSFORM(
-    DataType.STRUCT([
-      DataType.UINT,
-      DataType.PLACEHOLDER(length => DataType.ABBREV_ARRAY(DataType.UINT, length))
-    ], function (members, dt) {
-      if (members.length === 1) {
-        return dt.resolve(members[0]);
-      }
-    }),
-    value => value[1]
-  )),
-  MaxClassAttribute: new CIPAttribute.Class(6, 'Max Class Attribute ID', DataType.UINT),
-  MaxInstanceAttribute: new CIPAttribute.Class(7, 'Max Instance Attribute ID', DataType.UINT),
-});
+// const CommonClassAttribute = Object.freeze({
+//   Revision: new CIPAttribute.Class(1, 'Revision', DataType.UINT),
+//   MaxInstance: new CIPAttribute.Class(2, 'Max Instance ID', DataType.UINT),
+//   NumberOfInstances: new CIPAttribute.Class(3, 'Number Of Instances', DataType.UINT),
+//   OptionalAttributeList: new CIPAttribute.Class(4, 'Optional Attribute List', DataType.TRANSFORM(
+//     DataType.STRUCT([
+//       DataType.UINT,
+//       DataType.PLACEHOLDER(length => DataType.ABBREV_ARRAY(DataType.UINT, length))
+//     ], function (members, dt) {
+//       if (members.length === 1) {
+//         return dt.resolve(members[0]);
+//       }
+//     }),
+//     value => value[1]
+//   )),
+//   OptionalServiceList: new CIPAttribute.Class(4, 'Optional Service List', DataType.TRANSFORM(
+//     DataType.STRUCT([
+//       DataType.UINT,
+//       DataType.PLACEHOLDER(length => DataType.ABBREV_ARRAY(DataType.UINT, length))
+//     ], function (members, dt) {
+//       if (members.length === 1) {
+//         return dt.resolve(members[0]);
+//       }
+//     }),
+//     value => value[1]
+//   )),
+//   MaxClassAttribute: new CIPAttribute.Class(6, 'Max Class Attribute ID', DataType.UINT),
+//   MaxInstanceAttribute: new CIPAttribute.Class(7, 'Max Instance Attribute ID', DataType.UINT),
+// });
 
-const CommonClassAttributeGroup = new CIPFeatureGroup(Object.values(CommonClassAttribute));
+// const CommonClassAttributeGroup = new CIPFeatureGroup(Object.values(CommonClassAttribute));
 
 
-const CommonServices = Object.freeze({
-  GetAttributeSingle: function(attribute, instance = 0) {
-    let attributeID;
-    if (attribute instanceof CIPAttribute.Class) {
-      attributeID = ClassAttributeGroup.getCode(attribute) || CommonClassAttributeGroup.getCode(attribute);
-    } else if (attribute instanceof CIPAttribute.Instance) {
-      attributeID = InstanceAttributeGroup.getCode(attribute) || CommonClassAttributeGroup.getCode(attribute);
-    } else {
-      throw new Error(`Attribute must be a CIPClassAttribute or CIPInstanceAttribute`);
-    }
-    return new CIPRequest(
-      CommonServiceCodes.GetAttributeSingle,
-      EPath.Encode(true, [
-        new EPath.Segments.Logical.ClassID(classCode),
-        new EPath.Segments.Logical.InstanceID(instance),
-        new EPath.Segments.Logical.AttributeID(attributeID)
-      ]),
-      null,
-      (buffer, offset, cb) => {
-        return DecodeAttribute(buffer, offset, attribute, cb);
-      }
-    );
-  }
-});
+// const CommonServices = Object.freeze({
+//   GetAttributeSingle: function(attribute, instance = 0) {
+//     let attributeID;
+//     if (attribute instanceof CIPAttribute.Class) {
+//       attributeID = ClassAttributeGroup.getCode(attribute) || CommonClassAttributeGroup.getCode(attribute);
+//     } else if (attribute instanceof CIPAttribute.Instance) {
+//       attributeID = InstanceAttributeGroup.getCode(attribute) || CommonClassAttributeGroup.getCode(attribute);
+//     } else {
+//       throw new Error(`Attribute must be a CIPClassAttribute or CIPInstanceAttribute`);
+//     }
+//     return new CIPRequest(
+//       CommonServiceCodes.GetAttributeSingle,
+//       EPath.Encode(true, [
+//         new EPath.Segments.Logical.ClassID(classCode),
+//         new EPath.Segments.Logical.InstanceID(instance),
+//         new EPath.Segments.Logical.AttributeID(attributeID)
+//       ]),
+//       null,
+//       (buffer, offset, cb) => {
+//         return DecodeAttribute(buffer, offset, attribute, cb);
+//       }
+//     );
+//   }
+// });
 
 function CIPMetaObject(classCode, options) {
   options = options || {};
-  // const ClassAttributeGroup = options.ClassAttributeGroup || new CIPFeatureGroup([]);
-  // const InstanceAttributeGroup = options.InstanceAttributeGroup || new CIPFeatureGroup([]);
 
+  const CommonClassAttribute = Object.freeze({
+    Revision: new CIPAttribute.Class(1, 'Revision', DataType.UINT),
+    MaxInstance: new CIPAttribute.Class(2, 'Max Instance ID', DataType.UINT),
+    NumberOfInstances: new CIPAttribute.Class(3, 'Number Of Instances', DataType.UINT),
+    OptionalAttributeList: new CIPAttribute.Class(4, 'Optional Attribute List', DataType.TRANSFORM(
+      DataType.STRUCT([
+        DataType.UINT,
+        DataType.PLACEHOLDER(length => DataType.ABBREV_ARRAY(DataType.UINT, length))
+      ], function (members, dt) {
+        if (members.length === 1) {
+          return dt.resolve(members[0]);
+        }
+      }),
+      value => value[1]
+    )),
+    OptionalServiceList: new CIPAttribute.Class(4, 'Optional Service List', DataType.TRANSFORM(
+      DataType.STRUCT([
+        DataType.UINT,
+        DataType.PLACEHOLDER(length => DataType.ABBREV_ARRAY(DataType.UINT, length))
+      ], function (members, dt) {
+        if (members.length === 1) {
+          return dt.resolve(members[0]);
+        }
+      }),
+      value => value[1]
+    )),
+    MaxClassAttribute: new CIPAttribute.Class(6, 'Max Class Attribute ID', DataType.UINT),
+    MaxInstanceAttribute: new CIPAttribute.Class(7, 'Max Instance Attribute ID', DataType.UINT),
+  });
+
+  
+  const CommonClassAttributes = Object.values(CommonClassAttribute);
   const ClassAttributes = Array.isArray(options.ClassAttributes) ? options.ClassAttributes : Object.values(options.ClassAttributes || {});
   const InstanceAttributes = Array.isArray(options.InstanceAttributes) ? options.InstanceAttributes : Object.values(options.InstanceAttributes || {});
   
+  CommonClassAttributes.forEach(attribute => attribute.classCode = classCode);
   ClassAttributes.forEach(attribute => attribute.classCode = classCode);
   InstanceAttributes.forEach(attribute => attribute.classCode = classCode);
 
+  const CommonClassAttributeGroup = new CIPFeatureGroup(CommonClassAttributes);
   const ClassAttributeGroup = new CIPFeatureGroup(ClassAttributes);
   const InstanceAttributeGroup = new CIPFeatureGroup(InstanceAttributes);
   const GetAttributesAllInstanceAttributes = options.GetAttributesAllInstanceAttributes || [];
@@ -131,13 +163,16 @@ function CIPMetaObject(classCode, options) {
 
 
     static GetAttributeSingle(attribute, instance) {
-      if (instance == null) {
+      let attributeID;
+      if (attribute instanceof CIPAttribute.Class) {
         /** class attribute */
-        instance = 0;
-        attribute = ClassAttributeGroup.getCode(attribute) || CommonClassAttribute.getCode(attribute);
+        if (instance == null) {
+          instance = 0;
+        }
+        attributeID = ClassAttributeGroup.getCode(attribute) || CommonClassAttribute.getCode(attribute);
       } else {
         /** instance attribute */
-        attribute = InstanceAttributeGroup.getCode(attribute);
+        attributeID = InstanceAttributeGroup.getCode(attribute);
       }
 
       return new CIPRequest(
@@ -145,7 +180,7 @@ function CIPMetaObject(classCode, options) {
         EPath.Encode(true, [
           new EPath.Segments.Logical.ClassID(classCode),
           new EPath.Segments.Logical.InstanceID(instance),
-          new EPath.Segments.Logical.AttributeID(attribute)
+          new EPath.Segments.Logical.AttributeID(attributeID)
         ]),
         null,
         (buffer, offset, cb) => {
@@ -154,14 +189,6 @@ function CIPMetaObject(classCode, options) {
       );
     }
     
-
-    // static GetAttributeList(attributes, instance) {
-    //   Port.Run(Port.CommonServices)
-    // }
-
-    // static Run(service) {
-
-    // }
 
     static DecodeAttribute(buffer, offset, attribute, cb) {
       if (attribute instanceof CIPAttribute.Class) {
@@ -203,7 +230,7 @@ function CIPMetaObject(classCode, options) {
   }
 
   CIPObject.CommonClassAttribute = CommonClassAttribute;
-  CIPObject.CommonServices = CommonServices;
+  // CIPObject.CommonServices = CommonServices;
 
   return CIPObject;
 }
@@ -225,10 +252,7 @@ function DecodeAttribute(buffer, offset, attribute, cb) {
 
   if (typeof cb === 'function') {
     cb({
-      attribute: {
-        code: attribute.code,
-        name: attribute.name
-      },
+      attribute,
       value
     });
   }